backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
langchain
langchain-google-genai
azure-storage-blob
python-pptx
pytesseract
langchain-astradb
pyjwt
cryptography
requests
langchain-community
pillow
PyJWT
